--- Player_Data.dat
@@ -1,3 +1,0 @@
-MujidLee%231982	1588308778000
-Bow%2311354	1588308778000
-Raj752%231280	1588480594000
