+++ app/api/.flaskenv
@@ -1,0 +1,2 @@
+FLASK_APP = api.py
+FLASK_ENV = development
