+++ app/api/templates/login_page.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <link rel="stylesheet" type="text/css" href="static/css/login.css">
+    </head>
+    <script>
+     function showPass(){
+         var x = document.getElementById("myPass");
+         if (x.type === "password"){
+             x.type="text";
+         }
+         else {
+             x.type = "password";
+         }
+     }
+    </script>
+    <body>
+        <div class="login">
+            <h1>Call of Duty Login</h1>
+            <form method="POST" action="{{url_for('login_info')}}">
+                <p><input type="text" name="user" value="" placeholder="CoD Email"></p>
+                <p><input type="password" name="pass" id="myPass" value="" placeholder="Password"></p>
+                <p><input type="checkbox" onclick="showPass()">Show Password</p>
+                <p class="submit"><input type="submit" name="commit" value="Login"></p>
+            </form>
+        </div>
+    </body>
+</html>
