+++ requirements.txt
@@ -1,6 +1,16 @@
-certifi==2020.4.5.1
-chardet==3.0.4
-idna==2.9
-lxml==4.5.0
-requests==2.23.0
-urllib3==1.25.9
+appdirs==1.4.3
+autoflake==1.3.1
+distlib==0.3.0
+entrypoints==0.3
+filelock==3.0.12
+flake8==3.7.9
+importlib-metadata==1.6.0
+jedi==0.17.0
+mccabe==0.6.1
+parso==0.7.0
+pycodestyle==2.5.0
+pyflakes==2.1.1
+six==1.14.0
+virtualenv==20.0.18
+yapf==0.30.0
+zipp==3.1.0
