--- templates/login_page.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <link rel="stylesheet" type="text/css" href="static/css/login.css">
-    </head>
-    <body>
-        <div class="login">
-            <h1>Call of Duty Login</h1>
-            <form method="post" action="">
-                <p><input type="text" name="login" value="" placeholder="CoD Email"></p>
-                <p><input type="text" name="login" value="" placeholder="Password"></p>
-                <p class="submit"><input type="submit" name="commit" value="Login"></p>
-            </form>
-        </div>
-    </body>
-</html>
